# app/Http/Controllers/Api/V1/Auth/MeController.php
<?php

namespace App\Http\Controllers\Api\V1\Auth;

use App\Http\Controllers\Api\ApiBaseController;
use App\Http\Controllers\Controller;
use App\Http\Resources\UserResource;
use Illuminate\Http\Request;

class MeController extends ApiBaseController
{
     /**
     * @OA\Get(
     *     path="/auth/me",
     *     tags={"Auth"},
     *     summary="Get current authenticated user",
     *     description="Returns current user details and token abilities. Requires authentication.",
     *     security={{"bearerAuth":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="User details retrieved",
     *         @OA\JsonContent(
     *             @OA\Property(property="user", type="object",
     *                 @OA\Property(property="id", type="integer", example=1),
     *                 @OA\Property(property="name", type="string", example="Ahmed Ali"),
     *                 @OA\Property(property="email", type="string", example="ahmed@example.com")
     *             ),
     *         )
     *     ),
     *     @OA\Response(response=401, description="Unauthenticated")
     * )
     */
    public function __invoke()
    {
        return response()->json([
            'user' => new UserResource($this->authUser),
        ]);
    }
}